Extract conversion methods to Convert module
